Refactor: Extract ButtonClickHandler type
This doesn't really belong to Card - it's a general concept
