+++ src/clickHandler.d.ts
@@ -1,3 +1,4 @@
 import type { MouseEvent } from "react";
 
 export type ClickHandler<T> = (event: MouseEvent<T>) => void;
+export type ButtonClickHandler = ClickHandler<HTMLButtonElement>;
